Modify MathUtils methods sum() & avg() parameter: int[] arr -> double[] arr.

+++ src/muikku.app/MathUtils.java
@@ -45,7 +45,7 @@ class MathUtils {
    * @param arr array containing ints.
    * @return sum of all elements in an int array.
    */
-  public static int sum(int[] arr) {
+  public static int sum(double[] arr) {
     int sum = 0;
     for (int i = 0; i < arr.length; i++) {
       sum += arr[i];
@@ -81,7 +81,7 @@ class MathUtils {
    * @param arr array to find avg from
    * @return average number from an array of ints
    */
-  public static double avg(int[] arr) {
+  public static double avg(double[] arr) {
     int sum = sum(arr);
     int amount = arr.length;
     return (1.0 * sum) / amount;
